Adding GitHub actions
Also making sure that `run-concurrent.sh` returns a non-zero error code if one of the processes fails.
Also making C code run for longer since it's harder to reproduce the problem.

And finally mentioning that I saw at least once the problem also in C, so it seems to be an underlying problem.